+++ scripts/lichessTools-init.js
@@ -64,6 +64,9 @@ lichessTools.loadTool(LiChessTools.Tools.TvOptions);
 lichessTools.loadTool(LiChessTools.Tools.ShowPawnStructure);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerSnaps);
 lichessTools.loadTool(LiChessTools.Tools.MiniGameSize);
+lichessTools.loadTool(LiChessTools.Tools.ForceSf16);
+lichessTools.loadTool(LiChessTools.Tools.StudyLearnFromMistakes);
+lichessTools.loadTool(LiChessTools.Tools.PinOnMainPage);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);
 lichessTools.loadTool(LiChessTools.Tools.TrapValueCommand);