+++ tools/explorerEval/tool.js
@@ -161,6 +161,7 @@
     setCached404= (path)=>path?this._setCached404(path,this.cache404):false;
     getCached404= (path)=>path?this._getCached404(path,this.cache404):false;
     _setCached404= (path,node)=>{
+      if (node===true) return;
       const key=path?.slice(0,2);
       if (key) {
         let newNode=node[key];
@@ -332,7 +333,7 @@
       }
     };
 
-    CSP=true; // defaults to true until lichess allows chessdb.cn connections
+    CSP=true; // default to true until lichess CSP rules allow chessdb.cn
     secCheck=e=>{
       if (this.CSP) return;
       if (!e.blockedURI?.includes('chessdb.cn')) {