+++ tools/explorerGambits/gambits.js
@@ -2898,17 +2898,17 @@
       ]
     },
     "r1bqkbnrpppp1ppp2n54p33PP35N2PPP2PPPRNBQKB1RbKQkq-": {
-      "total": 7,
+      "total": 8,
       "moves": [
+        {
+          "uci": "e5d4",
+          "san": "exd4",
+          "nr": 7
+        },
         {
           "uci": "c6d4",
           "san": "Nxd4",
           "nr": 1
-        },
-        {
-          "uci": "e5d4",
-          "san": "exd4",
-          "nr": 6
         }
       ]
     },
@@ -3028,12 +3028,12 @@
       ]
     },
     "r1bqkbnrpppp1ppp2n54p34P35N2PPPP1PPPRNBQKB1RwKQkqd3": {
-      "total": 7,
+      "total": 8,
       "moves": [
         {
           "uci": "d2d4",
           "san": "d4",
-          "nr": 7
+          "nr": 8
         }
       ]
     },
@@ -3138,8 +3138,13 @@
       ]
     },
     "r1bqkbnrpppp1ppp2n583pP35N2PPP2PPPRNBQKB1RwKQkq-": {
-      "total": 6,
+      "total": 7,
       "moves": [
+        {
+          "uci": "f3g5",
+          "san": "Ng5",
+          "nr": 1
+        },
         {
           "uci": "f1b5",
           "san": "Bb5",
@@ -13578,12 +13583,12 @@
       ]
     },
     "rnbqkbnrpppp1ppp84p34P35N2PPPP1PPPRNBQKB1RbKQkq-": {
-      "total": 71,
+      "total": 72,
       "moves": [
         {
           "uci": "b8c6",
           "san": "Nc6",
-          "nr": 56
+          "nr": 57
         },
         {
           "uci": "g8f6",
@@ -13618,8 +13623,13 @@
       ]
     },
     "rnbqkbnrpppp1ppp84p34P38PPPP1PPPRNBQKBNRwKQkq-": {
-      "total": 112,
+      "total": 113,
       "moves": [
+        {
+          "uci": "g1f3",
+          "san": "Nf3",
+          "nr": 78
+        },
         {
           "uci": "f1b5",
           "san": "Bb5",
@@ -13634,11 +13644,6 @@
           "uci": "b1c3",
           "san": "Nc3",
           "nr": 14
-        },
-        {
-          "uci": "g1f3",
-          "san": "Nf3",
-          "nr": 77
         }
       ]
     },
@@ -14968,8 +14973,13 @@
       ]
     },
     "rnbqkbnrpppppppp884P38PPPP1PPPRNBQKBNRbKQkq-": {
-      "total": 314,
+      "total": 315,
       "moves": [
+        {
+          "uci": "e7e5",
+          "san": "e5",
+          "nr": 199
+        },
         {
           "uci": "g8f6",
           "san": "Nf6",
@@ -15004,11 +15014,6 @@
           "uci": "d7d6",
           "san": "d6",
           "nr": 2
-        },
-        {
-          "uci": "e7e5",
-          "san": "e5",
-          "nr": 198
         },
         {
           "uci": "e7e6",
@@ -15228,12 +15233,12 @@
       ]
     },
     "rnbqkbnrpppppppp8888PPPPPPPPRNBQKBNRwKQkq-": {
-      "total": 506,
+      "total": 507,
       "moves": [
         {
           "uci": "e2e4",
           "san": "e4",
-          "nr": 314
+          "nr": 315
         },
         {
           "uci": "b1c3",
@@ -21895,17 +21900,17 @@
       ]
     },
     "rnbqkb1rppp1pppp5n23P42P58PP1P1PPPRNBQKBNRbKQkq-": {
-      "total": 3,
+      "total": 4,
       "moves": [
+        {
+          "uci": "e7e6",
+          "san": "e6",
+          "nr": 3
+        },
         {
           "uci": "c7c6",
           "san": "c6",
           "nr": 1
-        },
-        {
-          "uci": "e7e6",
-          "san": "e6",
-          "nr": 2
         }
       ]
     },
@@ -22035,17 +22040,17 @@
       ]
     },
     "rnbqkb1rppp1pppp5n23P488PPPP1PPPRNBQKBNRwKQkq-": {
-      "total": 7,
+      "total": 8,
       "moves": [
+        {
+          "uci": "c2c4",
+          "san": "c4",
+          "nr": 4
+        },
         {
           "uci": "b1c3",
           "san": "Nc3",
           "nr": 1
-        },
-        {
-          "uci": "c2c4",
-          "san": "c4",
-          "nr": 3
         },
         {
           "uci": "d2d4",
@@ -22215,12 +22220,12 @@
       ]
     },
     "rnbqkb1rppp2ppp4pn23P42P58PP1P1PPPRNBQKBNRwKQkq-": {
-      "total": 1,
+      "total": 2,
       "moves": [
         {
           "uci": "d5e6",
           "san": "dxe6",
-          "nr": 1
+          "nr": 2
         }
       ]
     },
@@ -22265,8 +22270,13 @@
       ]
     },
     "rnbqkb1rppp2ppp4Pn282P58PP1P1PPPRNBQKBNRbKQkq-": {
-      "total": 1,
+      "total": 2,
       "moves": [
+        {
+          "uci": "b8c6",
+          "san": "Nc6",
+          "nr": 1
+        },
         {
           "uci": "f8c5",
           "san": "Bc5",
@@ -25130,8 +25140,13 @@
       ]
     },
     "rnbqkbnrppp1pppp83p44P38PPPP1PPPRNBQKBNRwKQkq-": {
-      "total": 15,
+      "total": 16,
       "moves": [
+        {
+          "uci": "e4d5",
+          "san": "exd5",
+          "nr": 13
+        },
         {
           "uci": "b1c3",
           "san": "Nc3",
@@ -25141,11 +25156,6 @@
           "uci": "g1f3",
           "san": "Nf3",
           "nr": 2
-        },
-        {
-          "uci": "e4d5",
-          "san": "exd5",
-          "nr": 12
         }
       ]
     },
@@ -25370,17 +25380,17 @@
       ]
     },
     "rnbqkbnrppp1pppp83P488PPPP1PPPRNBQKBNRbKQkq-": {
-      "total": 12,
+      "total": 13,
       "moves": [
+        {
+          "uci": "g8f6",
+          "san": "Nf6",
+          "nr": 8
+        },
         {
           "uci": "e7e5",
           "san": "e5",
           "nr": 1
-        },
-        {
-          "uci": "g8f6",
-          "san": "Nf6",
-          "nr": 7
         },
         {
           "uci": "c7c6",
@@ -27655,8 +27665,13 @@
       ]
     },
     "rnbqkbnrpppppppp884P38PPPP1PPPRNBQKBNRbKQkq-": {
-      "total": 201,
+      "total": 202,
       "moves": [
+        {
+          "uci": "d7d5",
+          "san": "d5",
+          "nr": 16
+        },
         {
           "uci": "b8c6",
           "san": "Nc6",
@@ -27686,11 +27701,6 @@
           "uci": "c7c6",
           "san": "c6",
           "nr": 1
-        },
-        {
-          "uci": "d7d5",
-          "san": "d5",
-          "nr": 15
         },
         {
           "uci": "d7d6",
@@ -27860,8 +27870,13 @@
       ]
     },
     "rnbqkbnrpppppppp8888PPPPPPPPRNBQKBNRwKQkq-": {
-      "total": 358,
+      "total": 359,
       "moves": [
+        {
+          "uci": "e2e4",
+          "san": "e4",
+          "nr": 202
+        },
         {
           "uci": "d2d4",
           "san": "d4",
@@ -27906,11 +27921,6 @@
           "uci": "d2d3",
           "san": "d3",
           "nr": 1
-        },
-        {
-          "uci": "e2e4",
-          "san": "e4",
-          "nr": 201
         },
         {
           "uci": "f2f4",