+++ HISTORY.md
@@ -1,5 +1,14 @@
 History of features added to LiChess Tools in time (reversed order)
 
+27 Dec
+  - Learn from your mistakes in Studies
+  - Pin study on home page
+
+26 Dec
+  - Slim arrows theme
+  - rewrote Move colorize theme as an option of the Highlight tool
+  - use Stockfish 16 on Brave browser
+
 21 Dec
   - collapsible Interactive Lesson edit controls